requirements.txt
alabaster==0.7.13
annotated-types==0.5.0
astunparse==1.6.3
autopep8==2.0.4
Babel==2.14.0
biopython==1.81
boto3==1.33.13
botocore==1.33.13
brotlipy==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.7.2
Cython==3.0.10
decorator==5.1.1
docutils==0.19
fair-esm==2.0.0
filelock==3.12.2
fonttools==4.25.0
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.7
imagesize==1.4.1
importlib-metadata==6.7.0
isort==5.11.5
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
lit==18.1.4
littleutils==0.2.2
lmdb==1.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
ninja==1.11.1.1
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
ogb==1.3.6
outdated==0.2.2
packaging==24.0
pandas==1.3.5
Pillow==9.4.0
ply==3.11
portalocker==2.7.0
protobuf==4.24.4
psutil==5.9.8
pycodestyle==2.10.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.2
PyQt5-sip==12.11.0
pytz==2023.3.post1
PyYAML==6.0.1
rdkit-pypi==2022.9.5
regex==2024.4.16
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.8.2
safetensors==0.4.3
seaborn==0.12.2
shellingham==1.5.4
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tape-proteins==0.5
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.66.4
transformers==4.30.2
typer==0.12.3
typing_extensions==4.7.1
tzdata==2024.1
urllib3==1.26.19
zipp==3.15.0
